requirements.txt
requests
replicate
evdev
Pillow
pydub
opencv-python
